+++ Assets/Scripts/Gameplay/UI/IPUIMediator.cs
@@ -41,12 +41,13 @@ namespace Unity.BossRoom.Gameplay.UI
 
         public IPHostingUI IPHostingUI => m_IPHostingUI;
 
-        IDisposable m_Subscription;
+        ISubscriber<ConnectStatus> m_ConnectStatusSubscriber;
 
         [Inject]
         void InjectDependencies(ISubscriber<ConnectStatus> connectStatusSubscriber)
         {
-            m_Subscription = connectStatusSubscriber.Subscribe(OnConnectStatusMessage);
+            m_ConnectStatusSubscriber = connectStatusSubscriber;
+            m_ConnectStatusSubscriber.Subscribe(OnConnectStatusMessage);
         }
 
         void Awake()
@@ -63,7 +64,7 @@ namespace Unity.BossRoom.Gameplay.UI
 
         void OnDestroy()
         {
-            m_Subscription.Dispose();
+            m_ConnectStatusSubscriber?.Unsubscribe(OnConnectStatusMessage);
         }
 
         void OnConnectStatusMessage(ConnectStatus connectStatus)