+++ Assets/Scripts/Gameplay/UI/Lobby/LobbyJoiningUI.cs
@@ -28,7 +28,8 @@ namespace Unity.BossRoom.Gameplay.UI
         IObjectResolver m_Container;
         LobbyUIMediator m_LobbyUIMediator;
         UpdateRunner m_UpdateRunner;
-        IDisposable m_Subscriptions;
+        ISubscriber<LobbyListFetchedMessage> m_LocalLobbiesRefreshedSub;
+
         List<LobbyListItemUI> m_LobbyListItems = new List<LobbyListItemUI>();
 
         void Awake()
@@ -40,13 +41,13 @@ namespace Unity.BossRoom.Gameplay.UI
         {
             if (m_UpdateRunner != null)
             {
-                m_UpdateRunner.Unsubscribe(PeriodicRefresh);
+                m_UpdateRunner?.Unsubscribe(PeriodicRefresh);
             }
         }
 
         void OnDestroy()
         {
-            m_Subscriptions?.Dispose();
+            m_LocalLobbiesRefreshedSub?.Unsubscribe(UpdateUI);
         }
 
         [Inject]
@@ -59,7 +60,8 @@ namespace Unity.BossRoom.Gameplay.UI
             m_Container = container;
             m_LobbyUIMediator = lobbyUIMediator;
             m_UpdateRunner = updateRunner;
-            m_Subscriptions = localLobbiesRefreshedSub.Subscribe(UpdateUI);
+            m_LocalLobbiesRefreshedSub = localLobbiesRefreshedSub;
+            m_LocalLobbiesRefreshedSub.Subscribe(UpdateUI);
         }
 
         /// <summary>